handle initial push notification in a custom hook
